// src/product/enums/product.enum.ts
export enum ProductMessages {
    NOT_FOUND = 'Product not found',
    ALREADY_EXISTS = 'Product already exists',
    CREATED = 'Product session created successfully',
    UPDATED = 'Product sesion updated successfully',
    DELETED = 'Product deleted successfully',
    GET_ALL_PRODUCTS = 'Get all products successfully',
    GET_PRODUCT = 'Get product successfully',
    GET_CLUB_PRODUCTS = 'Get club products successfully',
}